chore(deployment): update deployment scripts
Update deploy contracts script, update storage base url that includes
kms in the path for instance

Pay attention to setup wallets also, the account that triggers the key
gen and crs gen in simulator config should be the one who triggers
decryption in gateway or a permission errors occurs in KMS BC

Easy solution is to use the connector 1 account for both.

Better solution is to have dedicated account. However it is a bit
tedious to found exactly which mnemonic us used where.